requirements.txt
# Voice Agent Dependencies
livekit-agents>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-noise-cancellation>=0.2.0
python-dotenv>=1.0.0
